More updates in Proline-Core and ProlineWeb-Core to use long Ids .
refs #7653

+++ Proline-DAL/src/main/scala/fr/proline/core/om/provider/lcms/impl/MapAlignmentSetLoader.scala
@@ -31,7 +31,7 @@ class MapAlignmentSetLoader( val sqlExec: SQLQueryExecution ) {
     ()
   }
   
-  def getMapAlignmentSets( mapSetId: Int ): Array[MapAlignmentSet] = {
+  def getMapAlignmentSets( mapSetId: Long ): Array[MapAlignmentSet] = {
     
     var colNames: Seq[String] = null
