+++ Proline-OMP/src/main/scala/fr/proline/core/service/msq/ExportQuantRSM.scala
@@ -3,7 +3,7 @@ package fr.proline.core.service.msq
 import fr.proline.api.service.IService
 import fr.proline.repository.IDataStoreConnectorFactory
 
-class ExportQuantRSM( dbManager: IDataStoreConnectorFactory, quantRsmId: Int ) extends IService {
+class ExportQuantRSM( dbManager: IDataStoreConnectorFactory, quantRsmId: Long ) extends IService {
   
   def runService() = {
